+++ CHANGELOG.md
@@ -4,6 +4,8 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Event Catchup is now Observable using the jee event system
 
 ## [2.0.0-M7] - 2019-03-22
 ### Fixed